# openapi/panvex.yaml
...
openapi: 3.0.3
info:
  title: Panvex Control Plane API
  description: |
    Operator-facing HTTP API for the Panvex fleet management
    control-plane. Migrated incrementally from hand-rolled Zod
    schemas; consult `web/src/shared/api/schemas/` for endpoints not
    yet covered here.
  version: "1.0.0-pre"
  contact:
    name: Panvex maintainers
    url: https://github.com/lost-coder/panvex
servers:
  - url: /
    description: Same-origin (the panel serves both API and UI)
paths:
...
  /api/agents:
    get:
      operationId: listAgents
      summary: List enrolled agents in operator's fleet scope
      description: |
        Returns every agent the calling operator can see, filtered by
        their fleet-group scope (R-S-14). Each entry carries the live
        runtime snapshot the panel last received plus optional
        certificate recovery state.
      tags: [agents]
      responses:
        '200':
          description: Agent inventory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/agents/enrollment-tokens:
    get:
      operationId: listEnrollmentTokens
      summary: List enrollment tokens visible to the operator
      description: |
        Returns every enrollment token whose fleet-group sits within
        the caller's scope (R-S-14). Raw token values are masked —
        only the creation response surfaces the bearer secret. Use
        `handle` to revoke without exposing the raw value.
      tags: [agents, enrollment]
      responses:
        '200':
          description: Token list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentTokenList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createEnrollmentToken
      summary: Mint a new enrollment token
      description: |
        Issues a one-time bootstrap token for an agent to enroll into
        the named fleet group. The raw `value` is returned ONLY here;
        listing endpoints expose only a masked form afterwards.
      tags: [agents, enrollment]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnrollmentTokenRequest'
      responses:
        '201':
          description: Token created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEnrollmentTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/agents/enrollment-tokens/{value}/revoke:
    parameters:
      - name: value
        in: path
        required: true
        description: |
          Either the raw token value or the stable handle (SHA-256
          prefix) returned in the listing payload. Both forms resolve
          to the same persistent record.
        schema:
          type: string
    post:
      operationId: revokeEnrollmentToken
      summary: Revoke an enrollment token
      description: |
        Marks the token as revoked so subsequent bootstrap attempts
        fail. Idempotent — revoking an already-revoked token returns
        204 without an audit event.
      tags: [agents, enrollment]
      responses:
        '204':
          description: Token revoked (or already revoked).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Agent UUID.
        schema:
          type: string
    patch:
      operationId: renameAgent
      summary: Rename an agent's display node-name
      description: |
        Updates the operator-facing label only; the agent's UUID and
        certificate identity are unaffected. Idempotent (no audit
        event when the new name equals the existing one).
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameAgentRequest'
      responses:
        '200':
          description: Updated agent record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deregisterAgent
      summary: Deregister an agent
      description: |
        Terminates the agent's gRPC session, deletes its persistent
        records (instances, recovery grants), and adds the agent ID
        to the in-memory revocation list so a reconnect with valid
        mTLS material is rejected at Connect time.
      tags: [agents]
      responses:
        '204':
          description: Agent deregistered.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{id}/fleet-group:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateAgentFleetGroup
      summary: Reassign an agent to a different fleet group
      description: |
        Operator must be scoped over BOTH the current and target
        groups. No-op responses (existing == target) are returned
        with the current agent shape so the client can refresh
        without a special-cased branch.
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentFleetGroupRequest'
      responses:
        '200':
          description: Agent reassigned (or no-op).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{id}/update:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: dispatchAgentUpdate
      summary: Enqueue a self-update job for an agent
      description: |
        Schedules an `agent_self_update` job carrying the resolved
        binary asset metadata. The agent picks it up via its
        long-poll session.
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchAgentUpdateRequest'
      responses:
        '200':
          description: Update job dispatched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchAgentUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{id}/transport-mode:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateAgentTransportMode
      summary: Switch an agent between inbound and outbound transport
      description: |
        Persists the new transport mode in the DB, enqueues a
        `switch_transport_mode` job, and notifies the in-process
        transport manager so outbound supervisors are spawned or
        torn down immediately. `dial_address` is required when
        switching to outbound; the agent's listen-bind defaults
        to `:<port>` derived from `dial_address` when
        `listen_address` is omitted.
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentTransportModeRequest'
      responses:
        '204':
          description: Mode updated and job dispatched.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{id}/telemt/update-strategy:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getTelemtUpdateStrategy
      summary: Get an agent's Telemt update strategy and live probe
      description: |
        Returns the persisted update strategy for the agent (`null` when
        never configured) alongside its live-reported probe (`null` until
        the agent has sent at least one snapshot carrying it — see
        `Agent.telemt_update_probe`), so the dashboard can render "not
        configured yet, but here's what we detected" in one round trip.
      tags: [agents]
      responses:
        '200':
          description: Strategy and probe for the agent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemtUpdateStrategyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: putTelemtUpdateStrategy
      summary: Set an agent's Telemt update strategy
      description: |
        Persists how `telemt.update` jobs should apply on this agent:
        `mode` is one of `binary` (a supported process supervisor fronts
        telemt; a binary swap + supervised restart is safe), `docker`
        (updates must go through the image), or `none`. For
        `mode: binary`, `restart_spec` (validated with the same rules the
        agent applies) and an absolute `binary_path` are required.
        Admin-only.
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemtUpdateStrategy'
      responses:
        '204':
          description: Strategy saved.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{id}/telemt/update:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: dispatchTelemtUpdate
      summary: Enqueue a telemt.update job for an agent's managed Telemt install
      description: |
        Schedules a `telemt.update` job that downloads, verifies, and swaps
        the agent's MANAGED Telemt binary in place (distinct from
        `POST /api/agents/{id}/update`, which replaces the AGENT's own
        binary), then runs a supervised restart per the agent's configured
        update strategy (`PUT /api/agents/{id}/telemt/update-strategy`).
        `version` defaults to the panel's cached latest known Telemt
        release when omitted. A guard chain (see `DispatchTelemtUpdateError`)
        rejects the request with 409 before enqueueing when the strategy is
        unconfigured, does not support a binary swap, the agent's live probe
        reports the update as unavailable, or no version is known — checked
        in that order. Admin-only, same trust bar as the update-strategy
        endpoints.
      tags: [agents]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchTelemtUpdateRequest'
      responses:
        '202':
          description: Update job dispatched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchTelemtUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            The update cannot be dispatched right now — see `code` in the
            response body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchTelemtUpdateError'
...
  /api/agents/{id}/certificate-recovery-grants:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: createAgentCertificateRecoveryGrant
      summary: Allow an agent to recover an expired certificate
      description: |
        Admins issue a time-bounded grant so an agent whose mTLS
        certificate has expired can present a proof-of-possession
        and receive a fresh certificate without re-enrolling.
      tags: [agents, recovery]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCertificateRecoveryGrantRequest'
      responses:
        '201':
          description: Grant created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentCertificateRecoveryGrant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    AgentList:
      type: array
      description: List response from `GET /api/agents`.
      items:
        $ref: '#/components/schemas/Agent'
...
    EnrollmentTokenList:
      type: array
      items:
        $ref: '#/components/schemas/EnrollmentTokenListItem'
...
    CreateEnrollmentTokenRequest:
      type: object
      required: [ttl_seconds]
      properties:
        fleet_group_id:
          type: string
          description: |
            Either the canonical UUID or the friendly slug; empty
            string falls back to the default fleet group.
        ttl_seconds:
          type: integer
          minimum: 1
          description: Token lifetime in seconds.
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Caller is not authenticated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Caller is authenticated but lacks the required role/scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource does not exist or is out-of-scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: panvex_session
      description: |
        Server-issued session cookie set by `POST /api/login`. Marked
        Secure + HttpOnly + SameSite=Lax in production; the bare
        non-prefixed name above is the dev fallback for plain-HTTP
        loopback (the production runtime sets `__Host-panvex_session`
        instead — same auth flow, different cookie attribute set).
security:
  - sessionCookie: []